# Add boolean to turn off derivative calculations