--- a/src/numerics/include/metaphysicl/dualnumber.h
+++ b/src/numerics/include/metaphysicl/dualnumber.h
@@ -34,11 +34,14 @@
 
 namespace MetaPhysicL {
 
-static bool do_derivatives = true;
-
 template <typename T, typename D>
 class NotADuckDualNumber;
 
+// static member initialization
+template <typename T, typename D>
+bool DualNumber<T,D>::do_derivatives = true;
+
+// Member definitions
 template <typename T, typename D>
 inline
 T&
@@ -82,9 +85,35 @@ DualNumber<T,D>::operator=(const DualNumber<T2,D2> & dn)
 {
   _val = dn.value();
 
-  if (!::MetaPhysicL::do_derivatives)
+  if (!do_derivatives)
     return *this;
   _deriv = dn.derivatives();
+  return *this;
+}
+
+template <typename T, typename D>
+inline
+DualNumber<T,D> &
+DualNumber<T,D>::operator=(const DualNumber<T,D> & dn)
+{
+  _val = dn.value();
+
+  if (!do_derivatives)
+    return *this;
+  _deriv = dn.derivatives();
+  return *this;
+}
+
+template <typename T, typename D>
+inline
+DualNumber<T,D> &
+DualNumber<T,D>::operator=(DualNumber<T,D> && dn)
+{
+  _val = std::move(dn.value());
+
+  if (!do_derivatives)
+    return *this;
+  _deriv = std::move(dn.derivatives());
   return *this;
 }
 
@@ -95,18 +124,42 @@ DualNumber<T,D> &
 DualNumber<T,D>::operator=(const NotADuckDualNumber<T2,D2> & nd_dn)
 {
   _val = nd_dn.value();
-  if (!::MetaPhysicL::do_derivatives)
+
+  if (!do_derivatives)
     return *this;
   _deriv = nd_dn.derivatives();
   return *this;
 }
 
 template <typename T, typename D>
+inline
+DualNumber<T,D>::DualNumber(const DualNumber<T,D> & dn) :
+    _val(dn.value())
+{
+  if (!do_derivatives)
+    return;
+
+  _deriv = dn.derivatives();
+}
+
+template <typename T, typename D>
+inline
+DualNumber<T,D>::DualNumber(DualNumber<T,D> && dn) :
+    _val(std::move(dn.value()))
+{
+  if (!do_derivatives)
+    return;
+
+  _deriv = std::move(dn.derivatives());
+}
+
+template <typename T, typename D>
 template <typename T2, typename D2>
 inline
-DualNumber<T,D>::DualNumber(const DualNumberSurrogate<T2,D2> & dns) : _val(dns.value())
+DualNumber<T,D>::DualNumber(const DualNumberSurrogate<T2,D2> & dns) :
+    _val(dns.value())
 {
-  if (!::MetaPhysicL::do_derivatives)
+  if (!do_derivatives)
     return;
   auto size = dns.derivatives().size();
   for (decltype(size) i = 0; i < size; ++i)
@@ -120,7 +173,8 @@ DualNumber<T,D> &
 DualNumber<T,D>::operator=(const DualNumberSurrogate<T2,D2> & dns)
 {
   _val = dns.value();
-  if (!::MetaPhysicL::do_derivatives)
+
+  if (!do_derivatives)
     return *this;
   auto size = dns.derivatives().size();
   for (decltype(size) i = 0; i < size; ++i)
@@ -135,7 +189,7 @@ DualNumber<T,D> &
 DualNumber<T,D>::operator=(const T2 & scalar)
 {
   _val = scalar;
-  if (!::MetaPhysicL::do_derivatives)
+  if (!do_derivatives)
     return *this;
   _deriv = 0;
   return *this;
@@ -148,16 +202,22 @@ template <typename T, typename D>
 template <typename T2>
 inline
 DualNumber<T,D>::DualNumber(const T2& val) :
-  _val  (DualNumberConstructor<T,D>::value(val)),
-  _deriv(DualNumberConstructor<T,D>::deriv(val)) {}
+  _val  (DualNumberConstructor<T,D>::value(val))
+{
+  if (do_derivatives)
+    _deriv = DualNumberConstructor<T,D>::deriv(val);
+}
 
 template <typename T, typename D>
 template <typename T2, typename D2>
 inline
 DualNumber<T,D>::DualNumber(const T2& val,
                             const D2& deriv) :
-  _val  (DualNumberConstructor<T,D>::value(val,deriv)),
-  _deriv(DualNumberConstructor<T,D>::deriv(val,deriv)) {}
+  _val  (DualNumberConstructor<T,D>::value(val,deriv))
+{
+  if (do_derivatives)
+    _deriv = DualNumberConstructor<T,D>::deriv(val,deriv);
+}
 
 // Some helpers for reducing temporary creation and memset, memcpy calls
 
@@ -170,7 +230,8 @@ template <typename T,
 inline void
 derivative_multiply_helper(DualNumber<T, D<N, DT>> & out, const DualNumber<T, D<N, DT>> & in)
 {
-  if (!::MetaPhysicL::do_derivatives)
+  // do_derivatives is a static member so only need to check one object
+  if (!out.do_derivatives)
     return;
   auto & din = in.derivatives();
   auto & dout = out.derivatives();
@@ -185,7 +246,8 @@ template <typename T, typename D>
 inline void
 derivative_multiply_helper(DualNumber<T, D> & out, const DualNumber<T, D> & in)
 {
-  if (!::MetaPhysicL::do_derivatives)
+  // do_derivatives is a static member so only need to check one object
+  if (!out.do_derivatives)
     return;
   if (&in == &out)
     out.derivatives() = out.derivatives() * in.value() + out.value() * in.derivatives();
@@ -200,7 +262,8 @@ template <typename T, typename D, typename T2, typename D2>
 inline void
 derivative_multiply_helper(DualNumber<T, D> & out, const DualNumber<T2, D2> & in)
 {
-  if (!::MetaPhysicL::do_derivatives)
+  // Potentially different classes so need to check both
+  if (!(out.do_derivatives || in.do_derivatives))
     return;
   out.derivatives() *= in.value();
   out.derivatives() += out.value() * in.derivatives();
@@ -215,7 +278,8 @@ template <typename T,
 inline void
 derivative_division_helper(DualNumber<T, D<N, DT>> & out, const DualNumber<T, D<N, DT>> & in)
 {
-  if (!::MetaPhysicL::do_derivatives)
+  // do_derivatives is a static member so only need to check one object
+  if (!out.do_derivatives)
     return;
   auto & din = in.derivatives();
   auto & dout = out.derivatives();
@@ -230,7 +294,8 @@ template <typename T, typename D>
 inline void
 derivative_division_helper(DualNumber<T, D> & out, const DualNumber<T, D> & in)
 {
-  if (!::MetaPhysicL::do_derivatives)
+  // do_derivatives is a static member so only need to check one object
+  if (!out.do_derivatives)
     return;
   if (&in == &out)
     out.derivatives() =
@@ -246,7 +311,8 @@ template <typename T, typename D, typename T2, typename D2>
 inline void
 derivative_division_helper(DualNumber<T, D> & out, const DualNumber<T2, D2> & in)
 {
-  if (!::MetaPhysicL::do_derivatives)
+  // Potentially different classes so need to check both
+  if (!(out.do_derivatives || in.do_derivatives))
     return;
   out.derivatives() /= in.value();
   out.derivatives() -= out.value()/(in.value()*in.value()) * in.derivatives();
@@ -267,7 +333,10 @@ inline \
 DualNumber<T,D>& \
 DualNumber<T,D>::operator opname##= (const T2& in) \
 { \
-  simplecalc; \
+  if (do_derivatives) \
+  { \
+    simplecalc; \
+  } \
   this->value() opname##= in; \
   return *this; \
 } \
@@ -278,7 +347,10 @@ inline \
 DualNumber<T,D>& \
 DualNumber<T,D>::operator opname##= (const DualNumber<T2,D2>& in) \
 { \
-  dualcalc; \
+  if (do_derivatives) \
+  { \
+    dualcalc; \
+  } \
   this->value() opname##= in.value(); \
   return *this; \
 } \
@@ -289,7 +361,10 @@ inline \
 DualNumber<T,D> & \
 DualNumber<T,D>::operator opname##= (const NotADuckDualNumber<T2,D2>& in) \
 { \
-  dualcalc; \
+  if (do_derivatives) \
+  { \
+    dualcalc; \
+  } \
   this->value() opname##= in.value(); \
   return *this; \
 } \
@@ -457,7 +532,7 @@ DualNumber<T,D> funcname (const DualNumber<T,D> & in) \
 { \
   DualNumber<T,D> returnval = in; \
   T funcval = std::funcname(in.value()); \
-  if (::MetaPhysicL::do_derivatives) \
+  if (in.do_derivatives) \
   { \
     precalc; \
     returnval.derivatives() *= derivative; \
@@ -471,7 +546,7 @@ inline \
 DualNumber<T,D> funcname (DualNumber<T,D> && in) \
 { \
   T funcval = std::funcname(in.value()); \
-  if (::MetaPhysicL::do_derivatives) \
+  if (in.do_derivatives) \
   { \
     precalc; \
     in.derivatives() *= derivative; \
@@ -503,7 +578,7 @@ inline \
 DualNumber<T,D> funcname (DualNumber<T,D> in) \
 { \
   T funcval = std::funcname(in.value()); \
-  if (::MetaPhysicL::do_derivatives) \
+  if (in.do_derivatives) \
   { \
     precalc; \
     in.derivatives() *= derivative; \
@@ -670,7 +745,7 @@ funcname (const DualNumber<T,D>& a, const DualNumber<T2,D2>& b) \
   typedef typename CompareTypes<DualNumber<T,D>,DualNumber<T2,D2> >::supertype type; \
  \
   TS funcval = std::funcname(a.value(), b.value()); \
-  if (!::MetaPhysicL::do_derivatives) \
+  if (!(a.do_derivatives || b.do_derivatives)) \
     return type(funcval, 0); \
   return type(funcval, derivative); \
 } \
@@ -681,7 +756,7 @@ DualNumber<T,D> \
 funcname (const DualNumber<T,D>& a, const DualNumber<T,D>& b) \
 { \
   T funcval = std::funcname(a.value(), b.value()); \
-  if (!::MetaPhysicL::do_derivatives) \
+  if (!a.do_derivatives) \
     return DualNumber<T,D>(funcval, 0); \
   return DualNumber<T,D>(funcval, derivative); \
 } \