--- a/src/numerics/include/metaphysicl/dualsparsenumbervector.h
+++ b/src/numerics/include/metaphysicl/dualsparsenumbervector.h
@@ -94,7 +94,7 @@ struct DivergenceVectorFunctor {
   DivergenceVectorFunctor
     (const T* in, typename DerivativeType<T>::type& out) :
       _in(in), _out(out) { out = 0; }
-    
+
   template <typename ValueType>
   inline void operator()() const {
     const unsigned int
@@ -146,9 +146,9 @@ gradient(const SparseNumberVector<T, IndexSet>& a)
 // DualNumber is subordinate to SparseNumberVector
 
 #define DualSparseNumberVector_comparisons(templatename) \
-template<typename T, typename T2, typename D, typename IndexSet, bool reverseorder> \
-struct templatename<SparseNumberVector<T2, IndexSet>, DualNumber<T, D>, reverseorder> { \
-  typedef SparseNumberVector<typename Symmetric##templatename<T2,DualNumber<T, D>,reverseorder>::supertype, IndexSet> supertype; \
+template<typename T, typename T2, typename D, typename IndexSet, bool asd, bool reverseorder> \
+struct templatename<SparseNumberVector<T2, IndexSet>, DualNumber<T, D, asd>, reverseorder> { \
+  typedef SparseNumberVector<typename Symmetric##templatename<T2,DualNumber<T, D,asd>,reverseorder>::supertype, IndexSet> supertype; \
 }
 
 DualSparseNumberVector_comparisons(CompareTypes);