I expect this one to be a little controversial. I'd suggest looking at idaholab/moose#14701 for context. We would like to remove templating from the MOOSE AD user experience without sacrificing our current performance. With a `static bool do_derivatives` member we can turn off derivative calculation except when we really want it (e.g. during Jacobian evaluation). The most annoying thing about this to me is "possibly uninitialized` warnings from the compiler, which unfortunately could in cases correspond to "definitely uninitialized". 

I will say that for the `powerlaw_sum` test case in https://github.com/idaholab/moose/pull/14701#issuecomment-587182782, this PR reduces the computation time for a non-sparse derivative container from 55 to 6 seconds, while for a sparse container it reduces from 13 seconds to 6 seconds.