@lindsayad 2020-02-20
FWIW all MOOSE test and modules tests in `oprof` and `dbg` modes...