@roystgnr 2020-02-20
"see if bool checks slow AD down" - well, do they?  ;-)  You mention computation time reduction when the bool is selectively turned off, but you don't mention whether there's any computation time increase when the bool is available but left on.  I'd especially be concerned about threaded cases, with every local variable having to check on a bool in shared cache from inside inner loops.

Although, either way, that's just empirical data from today's hardware, and I'd swear that over the last couple decades I've watched performance rules of thumb go through at least a couple swings from "branches don't even count as using CPU time" to "branches can throw away a whole pipeline and cost you countless clock cycles" and back again.  Could we make the question moot and avoid controversy by adding both the static bool member variable ```do_derivatives``` *and* a bool template parameter ```sometimes_skip_derivatives``` which defaults to false?  Then people who want runtime flexibility can still easily instantiate in such a way that they get it, but for users who don't, an if statement like ```if (!do_derivatives)``` becomes ```if (sometimes_skip_derivatives && !do_derivatives)```, and that compiles away to nothing in the ```sometimes_skip_derivatives==false``` case, and then we don't need to worry about what rules of thumb future hardware brings.