@lindsayad 2020-02-20
> "see if bool checks slow AD down" - well, do they? ;-) You mention computation time reduction when the bool is selectively turned off, but you don't mention whether there's any computation time increase when the bool is available but left on.

@rwcarlsen  told me he saw virtually no slow down testing the exact case that you mention: having the `bool` available but left on.



>  Could we make the question moot and avoid controversy by adding both the static bool member variable `do_derivatives` _and_ a bool template parameter `sometimes_skip_derivatives` which defaults to false? Then people who want runtime flexibility can still easily instantiate in such a way that they get it, but for users who don't, an if statement like `if (!do_derivatives)` becomes `if (sometimes_skip_derivatives && !do_derivatives)`, and that compiles away to nothing in the `sometimes_skip_derivatives==false` case, and then we don't need to worry about what rules of thumb future hardware brings.

I like this suggestion