@roystgnr 2020-02-20
> @rwcarlsen told me he saw virtually no slow down testing the exact case that you mention: having the bool available but left on.

Okay.  Then I'm fine merging this as-is if you prefer, but I'd be happier merging after we have simplified branching, and even the template parameter if you're up for that.