@lindsayad 2020-02-20
> Okay. Then I'm fine merging this as-is if you prefer, but I'd be happier merging after we have simplified branching, and even the template parameter if you're up for that.

Already working on this, so let's wait