@lindsayad 2020-02-21
Ok, I think this is ready to review/merge