@lindsayad 2020-02-21
Summarizing some performance results for a PJFNK incompressible Navier-stokes problem with 202 residual evaluations and 3 Jacobian evaluations (also tagging @fdkong @rwcarlsen):

### non-sparse configurations (`D = NumberArray`)

- current MOOSE, current MetaPhysicL `master`: 57 seconds
- current MOOSE, this PR with `allow_skipping_deriviatives = false`: 56 seconds
- idaholab/moose#14701, this PR with `allow_skipping_derivatives = true` and only `do_derivatives` during Jacobian evaluation: 91 seconds
- idaholab/moose#14701, this PR with `allow_skipping_derivatives = true` and `do_derivatives` all the time: 584 seconds

### sparse configurations (`D = SemiDynamicSparseNumberArray`)

- current MOOSE, current MetaPhysicL `master`: 53 seconds
- current MOOSE, this PR with `allow_skipping_derivatives = false`: 53 seconds
- idaholab/moose#14701, this PR with `allow_skipping_derivatives = true` and only `do_derivatives` during Jacobian evaluation: 80 seconds
- idaholab/moose#14701, this PR with `allow_skipping_derivatives = true` and `do_derivatives` all the time: 138 seconds

### Summary

- `allow_skipping_derivatives = false` template parameter allows us to maintain optimal performance
- The `do_derivatives` static member allows us to keep under a 2x slow-down for both container types if we decide to move to a non-templated AD approach
    - The under 2x slow-down is at least valid for the case studied here, where are using a fairly expensive preconditioner (`-pc_type ilu -pc_factor_levels 8`). But PJFNK is also kind of a worst case scenario for AD anyways...